# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "MJ Ma",
            "url": "https://www.linkedin.com/in/mj-profile/",
            "email": "m.mjw.ma@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/shellingford37/jump-jump/blob/master/LICENSE"
        }
    },
    "paths": {
...
        "/short-link/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "短链接列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "短链接"
                ],
                "summary": "短链接列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ListShortLinksAPIResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": ""
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建短链接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "短链接"
                ],
                "summary": "创建短链接",
                "parameters": [
                    {
                        "description": "创建短链接请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateShortLinkAPIRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/CreateShortLinkAPIResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": ""
                    }
                }
            }
        },
        "/short-link/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取系统配置信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "短链接"
                ],
                "summary": "获取指定 ID 短链接",
                "parameters": [
                    {
                        "type": "string",
                        "description": "短链接 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/GetShortLinkAPIResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除短链接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "短链接"
                ],
                "summary": "删除短链接",
                "parameters": [
                    {
                        "type": "string",
                        "description": "短链接 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "401": {
                        "description": ""
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新短链接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "短链接"
                ],
                "summary": "更新短链接",
                "parameters": [
                    {
                        "type": "string",
                        "description": "短链接 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新短链接请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateShortLinkAPIRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/UpdateShortLinkAPIResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": ""
                    }
                }
            }
        },
        "/short-link/{id}/data": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "可查询短链接某个日期范围内的访问数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "短链接"
                ],
                "summary": "短链接访问数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "短链接 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "开始日期 YYYY-mm-dd",
                        "name": "startDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "结束日期 YYYY-mm-dd",
                        "name": "endDate",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ShortLinkDataAPIResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "CreateShortLinkAPIRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "描述",
                    "type": "string",
                    "example": "Jump Jump project"
                },
                "id": {
                    "description": "只有管理员可以在创建的时候指定 ID",
                    "type": "string",
                    "format": "string",
                    "example": "RANDOM_ID"
                },
                "idLength": {
                    "description": "短链接 ID 长度",
                    "type": "integer",
                    "format": "int",
                    "example": 4
                },
                "isEnable": {
                    "description": "是否启用",
                    "type": "boolean",
                    "format": "boolean",
                    "example": true
                },
                "url": {
                    "description": "目标链接",
                    "type": "string",
                    "example": "https://github.com/shellingford37/jump-jump"
                }
            }
        },
        "CreateShortLinkAPIResponseData": {
            "type": "object",
            "properties": {
                "shortLink": {
                    "$ref": "#/definitions/ShortLinkData"
                }
            }
        },
...
        "GetShortLinkAPIResponseData": {
            "type": "object",
            "properties": {
                "shortLink": {
                    "$ref": "#/definitions/ShortLinkData"
                }
            }
        },
...
        "ListShortLinksAPIResponseData": {
            "type": "object",
            "properties": {
                "shortLinks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ShortLinkData"
                    }
                },
                "total": {
                    "type": "integer",
                    "format": "10",
                    "example": 10
                }
            }
        },
...
        "RequestHistory": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "ua": {
                    "type": "string"
                },
                "url": {
                    "description": "由于短链接的目标连接可能会被修改，可以在访问历史记录中记录一下当前的目标连接",
                    "type": "string"
                }
            }
        },
        "Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int",
                    "default": 0,
                    "example": 0
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string",
                    "default": "ok",
                    "example": "ok"
                }
            }
        },
        "ShortLinkData": {
            "type": "object",
            "properties": {
                "createTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "createdBy": {
                    "description": "创建者",
                    "type": "string",
                    "format": "string",
                    "example": "admin"
                },
                "description": {
                    "description": "描述",
                    "type": "string",
                    "format": "string",
                    "example": "Jump Jump project"
                },
                "id": {
                    "type": "string",
                    "format": "string",
                    "example": "RANDOM_ID"
                },
                "isEnable": {
                    "description": "是否启用",
                    "type": "boolean",
                    "format": "boolean",
                    "example": true
                },
                "updateTime": {
                    "description": "最后更新时间",
                    "type": "string"
                },
                "url": {
                    "description": "目标链接",
                    "type": "string",
                    "format": "string",
                    "example": "https://github.com/shellingford37/jump-jump"
                }
            }
        },
        "ShortLinkDataAPIResponseData": {
            "type": "object",
            "properties": {
                "histories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/RequestHistory"
                    }
                }
            }
        },
...
        "UpdateShortLinkAPIRequest": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "description": {
                    "description": "描述",
                    "type": "string",
                    "example": "Jump Jump project"
                },
                "isEnable": {
                    "description": "是否启用",
                    "type": "boolean",
                    "format": "boolean",
                    "example": true
                },
                "url": {
                    "description": "目标链接",
                    "type": "string",
                    "example": "https://github.com/shellingford37/jump-jump"
                }
            }
        },
        "UpdateShortLinkAPIResponseData": {
            "type": "object",
            "properties": {
                "shortLink": {
                    "$ref": "#/definitions/ShortLinkData"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}